fix(app): satisfy current lint rules
