+++ packages/app/src/docker-git/menu-create-errors.ts
@@ -1,0 +1,7 @@
+import type { ParseError } from "./frontend-lib/core/domain.js"
+
+export const createParseError = (reason: string): ParseError => ({
+  _tag: "InvalidOption",
+  option: "create",
+  reason
+})
