+++ packages/app/src/web/app-ready-ssh-link-hook.ts
@@ -6,15 +6,28 @@ import type { BrowserActionContext } from "./actions-shared.js"
 import type { TerminalSession } from "./api-types.js"
 import { loadProjectTerminalWorkspace, loadTerminalSessionById } from "./api.js"
 import type { DashboardData } from "./api.js"
+import {
+  type DashboardProject,
+  findLocalTerminalSession,
+  findProjectBySshToken,
+  type ProjectLookupResult,
+  readSshLinkRequestFromHref,
+  resolveMissingSshSessionFallbackPath,
+  selectLocalProjectTerminal,
+  selectWorkspaceTerminalSession,
+  type SshLinkRequest,
+  sshLinkRequestKey
+} from "./app-ready-ssh-link-core.js"
 import { browserMenuIndex } from "./menu.js"
 import { projectPickerScreen } from "./screen.js"
-import { terminalSessionId, terminalSessionsForProject } from "./terminal-state.js"
-import {
-  type ActiveTerminalSession,
-  buildProjectActiveTerminalSession,
-  projectSshRoutePath,
-  terminalSessionRoutePath
-} from "./terminal.js"
+import { terminalSessionId } from "./terminal-state.js"
+import { type ActiveTerminalSession, buildProjectActiveTerminalSession, projectSshRoutePath } from "./terminal.js"
+
+export {
+  readSshLinkRequestFromHref,
+  resolveMissingSshSessionFallbackPath,
+  selectWorkspaceTerminalSession
+} from "./app-ready-ssh-link-core.js"
 
 type SshLinkArgs = {
   readonly actionContext: BrowserActionContext
@@ -27,15 +40,8 @@ type SshLinkArgs = {
   readonly terminalSessions: ReadonlyArray<ActiveTerminalSession>
 }
 
-const sshPathPrefix = "/ssh/"
 type ConnectTimerRef = { current: ReturnType<typeof globalThis.setTimeout> | null }
 type SshTokenRef = { current: string | null }
-type DashboardProject = DashboardData["projects"][number]
-type SessionLookupResult = { readonly sessionId: string }
-type ProjectLookupResult = { readonly terminalId?: string | undefined; readonly token: string }
-export type SshLinkRequest =
-  | ({ readonly kind: "project" } & ProjectLookupResult)
-  | ({ readonly kind: "session" } & SessionLookupResult)
 type SshLinkEffectArgs = Omit<SshLinkArgs, "dashboard"> & {
   readonly connectTimerRef: ConnectTimerRef
   readonly handledTokenRef: SshTokenRef
@@ -49,120 +55,12 @@ const clearConnectTimer = (connectTimerRef: ConnectTimerRef): void => {
   }
 }
 
-const decodePathTail = (value: string): string =>
-  value
-    .split("/")
-    .filter((segment) => segment.length > 0)
-    .map((segment) => decodeURIComponent(segment))
-    .join("/")
-    .trim()
-
-const readSessionPathRequest = (tail: string): SshLinkRequest | null => {
-  const sessionId = decodeURIComponent(tail.slice("session/".length).split("/")[0] ?? "").trim()
-  return sessionId.length === 0 ? null : { kind: "session", sessionId }
-}
-
-const readSshPathRequest = (url: URL): SshLinkRequest | null => {
-  if (!url.pathname.startsWith(sshPathPrefix)) {
-    return null
-  }
-  const tail = url.pathname.slice(sshPathPrefix.length)
-  if (tail.startsWith("session/")) {
-    return readSessionPathRequest(tail)
-  }
-  const decoded = decodePathTail(tail)
-  const terminalId = url.searchParams.get("terminal")?.trim() || url.searchParams.get("t")?.trim() || undefined
-  return decoded.length === 0 ? null : { kind: "project", terminalId, token: decoded }
-}
-
-const readSshQueryRequest = (url: URL): SshLinkRequest | null => {
-  const queryToken = url.searchParams.get("ssh")?.trim() ?? ""
-  const terminalId = url.searchParams.get("terminal")?.trim() || url.searchParams.get("t")?.trim() || undefined
-  return queryToken.length === 0 ? null : { kind: "project", terminalId, token: queryToken }
-}
-
-export const readSshLinkRequestFromHref = (href: string): SshLinkRequest | null => {
-  const url = new URL(href, "http://localhost")
-  return readSshPathRequest(url) ?? readSshQueryRequest(url)
-}
-
 const readSshLinkRequest = (): SshLinkRequest | null => readSshLinkRequestFromHref(globalThis.location.href)
-
-const findProjectBySshToken = (
-  projects: DashboardData["projects"],
-  token: string
-): DashboardProject | undefined =>
-  projects.find((candidate) => candidate.projectKey === token || candidate.id === token)
 
 const showProjectTerminalScreen = (actionContext: BrowserActionContext, projectId: string): void => {
   actionContext.setSelectedMenuIndex(browserMenuIndex("Select"))
   actionContext.setActiveScreen(projectPickerScreen())
   actionContext.setSelectedProjectId(projectId)
-}
-
-const findLocalTerminalSession = (
-  sessions: ReadonlyArray<ActiveTerminalSession>,
-  sessionId: string
-): ActiveTerminalSession | undefined => sessions.find((session) => terminalSessionId(session) === sessionId)
-
-const newestTerminalSession = <A extends { readonly createdAt: string; readonly status?: string }>(
-  sessions: ReadonlyArray<A>
-): A | null => {
-  const reusableSessions = sessions.filter((session) => session.status !== "failed")
-  const candidates = reusableSessions.length === 0 ? sessions : reusableSessions
-  return candidates.toSorted((left, right) => right.createdAt.localeCompare(left.createdAt))[0] ?? null
-}
-
-const selectByExactIdOrUniquePrefix = <A extends { readonly id: string }>(
-  sessions: ReadonlyArray<A>,
-  selector: string
-): A | null => {
-  const exact = sessions.find((session) => session.id === selector)
-  if (exact !== undefined) {
-    return exact
-  }
-  const matches = sessions.filter((session) => session.id.startsWith(selector))
-  return matches.length === 1 ? matches[0] ?? null : null
-}
-
-const selectLocalProjectTerminal = (
-  sessions: ReadonlyArray<ActiveTerminalSession>,
-  activeTerminalSessionId: string | null,
-  projectId: string,
-  terminalId: string | undefined
-): ActiveTerminalSession | null => {
-  const projectSessions = terminalSessionsForProject(sessions, projectId)
-  if (terminalId !== undefined) {
-    return selectByExactIdOrUniquePrefix(
-      projectSessions.map((session) => ({ ...session, id: terminalSessionId(session) })),
-      terminalId
-    )
-  }
-  const active = projectSessions.find((session) => terminalSessionId(session) === activeTerminalSessionId)
-  if (active !== undefined) {
-    return active
-  }
-  const newest = newestTerminalSession(projectSessions.map((session) => session.session))
-  return newest === null
-    ? null
-    : projectSessions.find((session) => terminalSessionId(session) === newest.id) ?? null
-}
-
-export const selectWorkspaceTerminalSession = (
-  sessions: ReadonlyArray<TerminalSession>,
-  activeSessionId: string | null,
-  terminalId?: string
-): TerminalSession | null => {
-  if (terminalId !== undefined) {
-    return selectByExactIdOrUniquePrefix(sessions, terminalId)
-  }
-  if (activeSessionId !== null) {
-    const active = sessions.find((session) => session.id === activeSessionId)
-    if (active !== undefined) {
-      return active
-    }
-  }
-  return newestTerminalSession(sessions)
 }
 
 const buildProjectTerminalSession = (
@@ -193,23 +91,6 @@ const attachProjectWorkspaceSessions = (
   }
   args.addTerminalSession(buildProjectTerminalSession(args, project, selectedSession))
   args.selectTerminalSession(selectedSession.id)
-}
-
-const sshLinkRequestKey = (request: SshLinkRequest): string =>
-  request.kind === "session"
-    ? `session:${request.sessionId}`
-    : `project:${request.token}:${request.terminalId ?? ""}`
-
-export const resolveMissingSshSessionFallbackPath = (
-  href: string,
-  sessionId: string,
-  error: string
-): string | null => {
-  if (!error.includes("HTTP 404")) {
-    return null
-  }
-  const url = new URL(href, "http://localhost")
-  return url.pathname === terminalSessionRoutePath(sessionId) ? "/menu/select" : null
 }
 
 const scheduleTerminalSessionAttach = (args: SshLinkEffectArgs, sessionId: string): void => {
