+++ packages/app/src/web/panel-project-details.tsx
@@ -1,6 +1,5 @@
-import { Effect, Match } from "effect"
+import { Match } from "effect"
 import type { JSX } from "react"
-import { useEffect, useState } from "react"
 
 import {
   buildSelectDetailsModel,
@@ -11,20 +10,24 @@ import {
 } from "../docker-git/menu-select-presenter.js"
 import { Box, Text } from "../ui/primitives.js"
 import { HelpLines } from "../ui/shared.js"
-import { loadProjectTerminalSessions, type ProjectDetails, type ProjectSummary, type TerminalSession } from "./api.js"
+import type { ProjectDetails, ProjectSummary } from "./api.js"
 import type { BrowserMenuTag } from "./menu.js"
-import { terminalTitleById } from "./terminal.js"
+import { ProjectTerminalSessionsForProject } from "./panel-project-terminal-sessions.js"
 
-type ProjectTerminalSessionsState = {
-  readonly error: string | null
-  readonly loading: boolean
-  readonly sessions: ReadonlyArray<TerminalSession>
-}
-
-const emptyProjectTerminalSessionsState: ProjectTerminalSessionsState = {
-  error: null,
-  loading: false,
-  sessions: []
+type SelectPanelProps = {
+  readonly currentMenu: BrowserMenuTag
+  readonly dashboardRefreshTick: number
+  readonly onAttachProjectTerminalSession: (
+    projectId: string,
+    projectKey: string,
+    projectDisplayName: string,
+    sessionId: string
+  ) => void
+  readonly onKillProjectTerminalSession: (projectId: string, projectKey: string, sessionId: string) => void
+  readonly onOpenProjectTerminalById: (projectId: string, projectKey?: string) => void
+  readonly project: ProjectDetails | null
+  readonly projectNavigationArmed: boolean
+  readonly selectedProjectSummary: ProjectSummary | undefined
 }
 
 const toSelectPurpose = (menu: BrowserMenuTag): SelectPurpose | null =>
@@ -97,146 +100,6 @@ const renderMissingProject = (purpose: SelectPurpose): JSX.Element => (
   </Box>
 )
 
-const formatSessionTime = (value: string | undefined): string =>
-  value === undefined ? "pending" : value.replace("T", " ").replace(/\.000Z$/u, " UTC")
-
-const terminalStatusColor = (status: TerminalSession["status"]): string => {
-  if (status === "attached") {
-    return "#56f39a"
-  }
-  if (status === "failed") {
-    return "#ff8aa0"
-  }
-  if (status === "exited") {
-    return "#ffd166"
-  }
-  return "#8fd3ff"
-}
-
-const sortedTerminalSessions = (
-  sessions: ReadonlyArray<TerminalSession>
-): ReadonlyArray<TerminalSession> => sessions.toSorted((left, right) => right.createdAt.localeCompare(left.createdAt))
-
-const ProjectTerminalSessionRow = (
-  {
-    onAttachProjectTerminalSession,
-    onKillProjectTerminalSession,
-    projectId,
-    projectKey,
-    projectName,
-    session,
-    title
-  }: {
-    readonly onAttachProjectTerminalSession: (
-      projectId: string,
-      projectKey: string,
-      projectDisplayName: string,
-      sessionId: string
-    ) => void
-    readonly onKillProjectTerminalSession: (projectId: string, projectKey: string, sessionId: string) => void
-    readonly projectId: string
-    readonly projectKey: string
-    readonly projectName: string
-    readonly session: TerminalSession
-    readonly title: string
-  }
-): JSX.Element => (
-  <Box border={true} borderColor="#3a4652" flexDirection="column" gap={1} padding={1}>
-    <Box alignItems="center" flexWrap="wrap" gap={1} justifyContent="space-between">
-      <Text bold={true} fg="#d6e5f7">{title}</Text>
-      <Text fg={terminalStatusColor(session.status)}>
-        {session.status}
-        {typeof session.attachedClients === "number" ? ` • clients ${session.attachedClients}` : ""}
-      </Text>
-    </Box>
-    <Text fg="#8fa6c4" wrap="truncate">created: {formatSessionTime(session.createdAt)}</Text>
-    <Text fg="#8fa6c4" wrap="truncate">started: {formatSessionTime(session.startedAt)}</Text>
-    <Text fg="#8fa6c4" wrap="truncate">{session.sshCommand}</Text>
-    <Box alignItems="center" flexWrap="wrap" gap={1}>
-      <Box
-        onClick={() => {
-          onAttachProjectTerminalSession(projectId, projectKey, projectName, session.id)
-        }}
-        width="auto"
-      >
-        <Text bold={true} fg="#78f0a3">attach</Text>
-      </Box>
-      <Box
-        onClick={() => {
-          onKillProjectTerminalSession(projectId, projectKey, session.id)
-        }}
-        width="auto"
-      >
-        <Text bold={true} fg="#ff8aa0">kill</Text>
-      </Box>
-    </Box>
-  </Box>
-)
-
-const ProjectTerminalSessionsSection = (
-  {
-    onAttachProjectTerminalSession,
-    onKillProjectTerminalSession,
-    onOpenProjectTerminalById,
-    onRefresh,
-    project,
-    sessionsState
-  }: {
-    readonly onAttachProjectTerminalSession: (
-      projectId: string,
-      projectKey: string,
-      projectDisplayName: string,
-      sessionId: string
-    ) => void
-    readonly onKillProjectTerminalSession: (projectId: string, projectKey: string, sessionId: string) => void
-    readonly onOpenProjectTerminalById: (projectId: string, projectKey?: string) => void
-    readonly onRefresh: () => void
-    readonly project: ProjectDetails
-    readonly sessionsState: ProjectTerminalSessionsState
-  }
-): JSX.Element => (
-  <Box flexDirection="column" marginTop={1}>
-    <Box alignItems="center" flexWrap="wrap" gap={1} justifyContent="space-between">
-      <Text bold={true} fg="#8be9fd">Live terminal sessions</Text>
-      <Box alignItems="center" flexWrap="wrap" gap={1}>
-        <Box
-          onClick={() => {
-            onOpenProjectTerminalById(project.id, project.projectKey)
-          }}
-          width="auto"
-        >
-          <Text bold={true} fg="#78f0a3">new terminal</Text>
-        </Box>
-        <Box onClick={onRefresh} width="auto">
-          <Text bold={true} fg="#7fdfff">refresh</Text>
-        </Box>
-      </Box>
-    </Box>
-    {sessionsState.loading ? <Text fg="#8fa6c4" marginTop={1}>Loading terminal sessions...</Text> : null}
-    {sessionsState.error === null ? null : <Text fg="#ff8aa0" marginTop={1}>{sessionsState.error}</Text>}
-    {!sessionsState.loading && sessionsState.error === null && sessionsState.sessions.length === 0
-      ? <Text fg="#8fa6c4" marginTop={1}>No live SSH terminals. Start one with `new terminal`.</Text>
-      : null}
-    <Box flexDirection="column" gap={1} marginTop={1}>
-      {(() => {
-        const terminalLabels = terminalTitleById(sessionsState.sessions)
-        return sortedTerminalSessions(sessionsState.sessions).map((session) => (
-          <ProjectTerminalSessionRow
-            key={session.id}
-            onAttachProjectTerminalSession={onAttachProjectTerminalSession}
-            onKillProjectTerminalSession={onKillProjectTerminalSession}
-            projectId={project.id}
-            projectKey={project.projectKey}
-            projectName={project.displayName}
-            session={session}
-            title={terminalLabels.get(session.id) ?? "Terminal"}
-          />
-        ))
-      })()}
-    </Box>
-  </Box>
-)
-
 const renderDetailsPanel = (
   purpose: SelectPurpose,
   project: ProjectDetails,
@@ -271,67 +134,9 @@ export const SelectPanel = (
     project,
     projectNavigationArmed,
     selectedProjectSummary
-  }: {
-    readonly currentMenu: BrowserMenuTag
-    readonly dashboardRefreshTick: number
-    readonly onAttachProjectTerminalSession: (
-      projectId: string,
-      projectKey: string,
-      projectDisplayName: string,
-      sessionId: string
-    ) => void
-    readonly onKillProjectTerminalSession: (projectId: string, projectKey: string, sessionId: string) => void
-    readonly onOpenProjectTerminalById: (projectId: string, projectKey?: string) => void
-    readonly project: ProjectDetails | null
-    readonly projectNavigationArmed: boolean
-    readonly selectedProjectSummary: ProjectSummary | undefined
-  }
+  }: SelectPanelProps
 ): JSX.Element | null => {
-  const [refreshNonce, setRefreshNonce] = useState(0)
-  const [sessionsState, setSessionsState] = useState<ProjectTerminalSessionsState>(emptyProjectTerminalSessionsState)
   const selectedProjectKey = selectedProjectSummary?.projectKey ?? null
-
-  useEffect(() => {
-    if (currentMenu !== "Select" || !projectNavigationArmed || selectedProjectKey === null) {
-      setSessionsState(emptyProjectTerminalSessionsState)
-      return
-    }
-    let cancelled = false
-    setSessionsState((current) => ({
-      error: null,
-      loading: true,
-      sessions: current.sessions
-    }))
-    void Effect.runPromise(
-      loadProjectTerminalSessions(selectedProjectKey).pipe(
-        Effect.match({
-          onFailure: (error) => {
-            if (cancelled) {
-              return
-            }
-            setSessionsState({
-              error,
-              loading: false,
-              sessions: []
-            })
-          },
-          onSuccess: (sessions) => {
-            if (cancelled) {
-              return
-            }
-            setSessionsState({
-              error: null,
-              loading: false,
-              sessions
-            })
-          }
-        })
-      )
-    )
-    return () => {
-      cancelled = true
-    }
-  }, [currentMenu, dashboardRefreshTick, projectNavigationArmed, refreshNonce, selectedProjectKey])
 
   if (currentMenu !== "Select") {
     return null
@@ -345,15 +150,15 @@ export const SelectPanel = (
   return (
     <Box flexDirection="column">
       {renderDetailsPanel("Connect", project, selectedProjectSummary)}
-      <ProjectTerminalSessionsSection
+      <ProjectTerminalSessionsForProject
+        currentMenu={currentMenu}
+        dashboardRefreshTick={dashboardRefreshTick}
         onAttachProjectTerminalSession={onAttachProjectTerminalSession}
         onKillProjectTerminalSession={onKillProjectTerminalSession}
         onOpenProjectTerminalById={onOpenProjectTerminalById}
-        onRefresh={() => {
-          setRefreshNonce((value) => value + 1)
-        }}
         project={project}
-        sessionsState={sessionsState}
+        projectNavigationArmed={projectNavigationArmed}
+        selectedProjectKey={selectedProjectKey}
       />
     </Box>
   )
