+++ packages/app/src/web/terminal-panel-inline-images-runtime.ts
@@ -1,0 +1,209 @@
+import { FetchHttpClient, HttpClient } from "@effect/platform"
+import { Effect } from "effect"
+
+import { resolveTerminalImageFetchUrl } from "./terminal-image-url.js"
+import {
+  splitTerminalInlineImageOutput,
+  type TerminalInlineImageOutputSegment,
+  writeTerminalOutputSegment
+} from "./terminal-inline-images-core.js"
+import {
+  appendTerminalInlineImagePreview,
+  cachedTerminalInlineImageEntry,
+  cacheTerminalInlineImageBlob,
+  terminalInlineImageSpacer,
+  unavailableTerminalInlineImageEntry
+} from "./terminal-inline-images.js"
+import type { TerminalInlineImageEntry } from "./terminal-inline-images.js"
+import type { TerminalMessageHandlers } from "./terminal-panel-runtime-types.js"
+
+type TerminalInlineImageFetchError = {
+  readonly _tag: "TerminalInlineImageFetchError"
+  readonly message: string
+}
+
+const terminalImageEntry = (
+  handlers: TerminalMessageHandlers,
+  path: string
+): TerminalInlineImageEntry | null => {
+  const fetchUrl = resolveTerminalImageFetchUrl(handlers.session.websocketPath, path)
+  return cachedTerminalInlineImageEntry(handlers.lifecycle.inlineImageObjectUrls, path, fetchUrl)
+}
+
+const terminalInlineImageFetchError = (message: string): TerminalInlineImageFetchError => ({
+  _tag: "TerminalInlineImageFetchError",
+  message
+})
+
+const terminalInlineImageFetchHeaders: Readonly<Record<string, string>> = {
+  accept: "image/*",
+  "cache-control": "no-cache, no-store, max-age=0",
+  pragma: "no-cache"
+}
+
+const imageBlobFromArrayBuffer = (
+  buffer: ArrayBuffer,
+  mediaType: string | undefined
+): Blob => new Blob([buffer], mediaType === undefined ? {} : { type: mediaType })
+
+const fetchTerminalInlineImageBlob = (
+  fetchUrl: string
+): Effect.Effect<Blob, TerminalInlineImageFetchError> =>
+  Effect.gen(function*(_) {
+    const client = yield* _(HttpClient.HttpClient)
+    const response = yield* _(
+      client.get(fetchUrl, { headers: terminalInlineImageFetchHeaders }).pipe(
+        Effect.mapError(() => terminalInlineImageFetchError("Could not fetch terminal image."))
+      )
+    )
+    if (response.status >= 400) {
+      return yield* _(Effect.fail(terminalInlineImageFetchError(`Terminal image returned HTTP ${response.status}.`)))
+    }
+    const buffer = yield* _(
+      response.arrayBuffer.pipe(
+        Effect.mapError(() => terminalInlineImageFetchError("Could not read terminal image response."))
+      )
+    )
+    return imageBlobFromArrayBuffer(buffer, response.headers["content-type"])
+  }).pipe(Effect.provide(FetchHttpClient.layer))
+
+const loadTerminalImageEntry = (
+  handlers: TerminalMessageHandlers,
+  path: string,
+  onComplete: (entry: TerminalInlineImageEntry) => void
+): void => {
+  const fetchUrl = resolveTerminalImageFetchUrl(handlers.session.websocketPath, path)
+  const cached = cachedTerminalInlineImageEntry(handlers.lifecycle.inlineImageObjectUrls, path, fetchUrl)
+  if (cached !== null) {
+    onComplete(cached)
+    return
+  }
+  Effect.runFork(
+    fetchTerminalInlineImageBlob(fetchUrl).pipe(
+      Effect.match({
+        onFailure: () => unavailableTerminalInlineImageEntry(path, fetchUrl),
+        onSuccess: (blob) =>
+          handlers.lifecycle.disposed
+            ? null
+            : cacheTerminalInlineImageBlob(handlers.lifecycle.inlineImageObjectUrls, path, fetchUrl, blob)
+      }),
+      Effect.flatMap((entry) =>
+        Effect.sync(() => {
+          if (entry === null || handlers.lifecycle.disposed) {
+            return
+          }
+          onComplete(entry)
+        })
+      )
+    )
+  )
+}
+
+const writePreviewSpacer = (
+  handlers: TerminalMessageHandlers,
+  onComplete: () => void
+): void => {
+  handlers.terminal.write(terminalInlineImageSpacer, onComplete)
+}
+
+const writeInlineImagePreview = (
+  handlers: TerminalMessageHandlers,
+  path: string,
+  onComplete: () => void
+): void => {
+  const cached = terminalImageEntry(handlers, path)
+  if (cached !== null) {
+    writeInlineImagePreviewEntry(handlers, cached, onComplete)
+    return
+  }
+  loadTerminalImageEntry(handlers, path, (entry) => {
+    writeInlineImagePreviewEntry(handlers, entry, onComplete)
+  })
+}
+
+const writeInlineImagePreviewEntry = (
+  handlers: TerminalMessageHandlers,
+  entry: TerminalInlineImageEntry,
+  onComplete: () => void
+): void => {
+  const appended = appendTerminalInlineImagePreview(
+    handlers.terminal,
+    handlers.lifecycle,
+    entry
+  )
+  if (!appended) {
+    onComplete()
+    return
+  }
+  writePreviewSpacer(handlers, onComplete)
+}
+
+const writeInlineImagePreviews = (
+  handlers: TerminalMessageHandlers,
+  paths: ReadonlyArray<string>,
+  onComplete: () => void
+): void => {
+  let index = 0
+  const writeNext = (): void => {
+    const path = paths[index]
+    if (path === undefined) {
+      onComplete()
+      return
+    }
+    index += 1
+    writeInlineImagePreview(handlers, path, writeNext)
+  }
+  writeNext()
+}
+
+const writeLineBreakBeforePreview = (
+  handlers: TerminalMessageHandlers,
+  segment: TerminalInlineImageOutputSegment,
+  onComplete: () => void
+): void => {
+  if (segment.endedWithLineBreak) {
+    onComplete()
+    return
+  }
+  handlers.terminal.write("\r\n", onComplete)
+}
+
+const flushTerminalOutputQueue = (handlers: TerminalMessageHandlers): void => {
+  if (handlers.lifecycle.outputWriting || handlers.lifecycle.disposed) {
+    return
+  }
+  const segment = handlers.lifecycle.outputQueue.shift()
+  if (segment === undefined) {
+    return
+  }
+
+  handlers.lifecycle.outputWriting = true
+  writeTerminalOutputSegment({
+    inlineImagePreviewsEnabledRef: handlers.inlineImagePreviewsEnabledRef,
+    segment,
+    writer: {
+      writePreviewLineBreak: (outputSegment, onComplete) => {
+        writeLineBreakBeforePreview(handlers, outputSegment, onComplete)
+      },
+      writePreviews: (paths, onComplete) => {
+        writeInlineImagePreviews(handlers, paths, onComplete)
+      },
+      writeText: (text, onComplete) => {
+        handlers.terminal.write(text, onComplete)
+      }
+    }
+  }, () => {
+    handlers.lifecycle.outputWriting = false
+    flushTerminalOutputQueue(handlers)
+  })
+}
+
+export const enqueueTerminalOutput = (
+  handlers: TerminalMessageHandlers,
+  data: string
+): void => {
+  for (const segment of splitTerminalInlineImageOutput(data)) {
+    handlers.lifecycle.outputQueue.push(segment)
+  }
+  flushTerminalOutputQueue(handlers)
+}
