+++ packages/app/src/web/terminal-panel-runtime-core.ts
@@ -1,26 +1,10 @@
-import { FetchHttpClient, HttpClient } from "@effect/platform"
-import { Effect, Either } from "effect"
 import { Terminal } from "xterm"
 import { FitAddon } from "xterm-addon-fit"
 
-import { resolveTerminalImageFetchUrl } from "./terminal-image-url.js"
-import {
-  splitTerminalInlineImageOutput,
-  type TerminalInlineImageOutputSegment,
-  writeTerminalOutputSegment
-} from "./terminal-inline-images-core.js"
-import {
-  appendTerminalInlineImagePreview,
-  cachedTerminalInlineImageEntry,
-  cacheTerminalInlineImageBlob,
-  revokeTerminalInlineImageObjectUrlCache,
-  terminalInlineImageSpacer,
-  unavailableTerminalInlineImageEntry
-} from "./terminal-inline-images.js"
-import type { TerminalInlineImageEntry } from "./terminal-inline-images.js"
+import { enqueueTerminalOutput } from "./terminal-panel-inline-images-runtime.js"
 import { sendTerminalClientMessage } from "./terminal-panel-input.js"
+import { runOptionalTerminalOperation } from "./terminal-panel-optional-operation.js"
 import type {
-  TerminalCleanupArgs,
   TerminalExitInfo,
   TerminalInputController,
   TerminalLifecycleState,
@@ -34,28 +18,12 @@ import { installTerminalQuerySuppression, type TerminalQuerySuppressionOptions }
 import { resolveTerminalReconnectDelay, terminalReconnectGraceMs } from "./terminal-reconnect.js"
 import { parseTerminalServerMessage, resolveTerminalWebSocketUrl } from "./terminal.js"
 
+export { cleanupTerminalResources } from "./terminal-panel-cleanup-runtime.js"
 export { attachTerminalInput, isTerminalMouseReportInput } from "./terminal-panel-input.js"
 
 type TerminalRuntimeOptions = {
   readonly querySuppression?: TerminalQuerySuppressionOptions
 }
-
-type TerminalInlineImageFetchError = {
-  readonly _tag: "TerminalInlineImageFetchError"
-  readonly message: string
-}
-
-const runOptionalTerminalOperation = (operation: () => void): boolean =>
-  Either.isRight(
-    Effect.runSync(
-      Effect.either(
-        Effect.try({
-          try: operation,
-          catch: (error) => error
-        })
-      )
-    )
-  )
 
 export const createLifecycleState = (): TerminalLifecycleState => ({
   attachedOnce: false,
@@ -191,192 +159,6 @@ const endTerminalSession = (
   }
 }
 
-const terminalImageEntry = (
-  handlers: TerminalMessageHandlers,
-  path: string
-): TerminalInlineImageEntry | null => {
-  const fetchUrl = resolveTerminalImageFetchUrl(handlers.session.websocketPath, path)
-  return cachedTerminalInlineImageEntry(handlers.lifecycle.inlineImageObjectUrls, path, fetchUrl)
-}
-
-const terminalInlineImageFetchError = (message: string): TerminalInlineImageFetchError => ({
-  _tag: "TerminalInlineImageFetchError",
-  message
-})
-
-const terminalInlineImageFetchHeaders: Readonly<Record<string, string>> = {
-  accept: "image/*",
-  "cache-control": "no-cache, no-store, max-age=0",
-  pragma: "no-cache"
-}
-
-const imageBlobFromArrayBuffer = (
-  buffer: ArrayBuffer,
-  mediaType: string | undefined
-): Blob => new Blob([buffer], mediaType === undefined ? {} : { type: mediaType })
-
-const fetchTerminalInlineImageBlob = (
-  fetchUrl: string
-): Effect.Effect<Blob, TerminalInlineImageFetchError> =>
-  Effect.gen(function*(_) {
-    const client = yield* _(HttpClient.HttpClient)
-    const response = yield* _(
-      client.get(fetchUrl, { headers: terminalInlineImageFetchHeaders }).pipe(
-        Effect.mapError(() => terminalInlineImageFetchError("Could not fetch terminal image."))
-      )
-    )
-    if (response.status >= 400) {
-      return yield* _(Effect.fail(terminalInlineImageFetchError(`Terminal image returned HTTP ${response.status}.`)))
-    }
-    const buffer = yield* _(
-      response.arrayBuffer.pipe(
-        Effect.mapError(() => terminalInlineImageFetchError("Could not read terminal image response."))
-      )
-    )
-    return imageBlobFromArrayBuffer(buffer, response.headers["content-type"])
-  }).pipe(Effect.provide(FetchHttpClient.layer))
-
-const loadTerminalImageEntry = (
-  handlers: TerminalMessageHandlers,
-  path: string,
-  onComplete: (entry: TerminalInlineImageEntry) => void
-): void => {
-  const fetchUrl = resolveTerminalImageFetchUrl(handlers.session.websocketPath, path)
-  const cached = cachedTerminalInlineImageEntry(handlers.lifecycle.inlineImageObjectUrls, path, fetchUrl)
-  if (cached !== null) {
-    onComplete(cached)
-    return
-  }
-  Effect.runFork(
-    fetchTerminalInlineImageBlob(fetchUrl).pipe(
-      Effect.match({
-        onFailure: () => unavailableTerminalInlineImageEntry(path, fetchUrl),
-        onSuccess: (blob) =>
-          handlers.lifecycle.disposed
-            ? null
-            : cacheTerminalInlineImageBlob(handlers.lifecycle.inlineImageObjectUrls, path, fetchUrl, blob)
-      }),
-      Effect.flatMap((entry) =>
-        Effect.sync(() => {
-          if (entry === null || handlers.lifecycle.disposed) {
-            return
-          }
-          onComplete(entry)
-        })
-      )
-    )
-  )
-}
-
-const writePreviewSpacer = (
-  handlers: TerminalMessageHandlers,
-  onComplete: () => void
-): void => {
-  handlers.terminal.write(terminalInlineImageSpacer, onComplete)
-}
-
-const writeInlineImagePreview = (
-  handlers: TerminalMessageHandlers,
-  path: string,
-  onComplete: () => void
-): void => {
-  const cached = terminalImageEntry(handlers, path)
-  if (cached !== null) {
-    writeInlineImagePreviewEntry(handlers, cached, onComplete)
-    return
-  }
-  loadTerminalImageEntry(handlers, path, (entry) => {
-    writeInlineImagePreviewEntry(handlers, entry, onComplete)
-  })
-}
-
-const writeInlineImagePreviewEntry = (
-  handlers: TerminalMessageHandlers,
-  entry: TerminalInlineImageEntry,
-  onComplete: () => void
-): void => {
-  const appended = appendTerminalInlineImagePreview(
-    handlers.terminal,
-    handlers.lifecycle,
-    entry
-  )
-  if (!appended) {
-    onComplete()
-    return
-  }
-  writePreviewSpacer(handlers, onComplete)
-}
-
-const writeInlineImagePreviews = (
-  handlers: TerminalMessageHandlers,
-  paths: ReadonlyArray<string>,
-  onComplete: () => void
-): void => {
-  let index = 0
-  const writeNext = (): void => {
-    const path = paths[index]
-    if (path === undefined) {
-      onComplete()
-      return
-    }
-    index += 1
-    writeInlineImagePreview(handlers, path, writeNext)
-  }
-  writeNext()
-}
-
-const writeLineBreakBeforePreview = (
-  handlers: TerminalMessageHandlers,
-  segment: TerminalInlineImageOutputSegment,
-  onComplete: () => void
-): void => {
-  if (segment.endedWithLineBreak) {
-    onComplete()
-    return
-  }
-  handlers.terminal.write("\r\n", onComplete)
-}
-
-const flushTerminalOutputQueue = (handlers: TerminalMessageHandlers): void => {
-  if (handlers.lifecycle.outputWriting || handlers.lifecycle.disposed) {
-    return
-  }
-  const segment = handlers.lifecycle.outputQueue.shift()
-  if (segment === undefined) {
-    return
-  }
-
-  handlers.lifecycle.outputWriting = true
-  writeTerminalOutputSegment({
-    inlineImagePreviewsEnabledRef: handlers.inlineImagePreviewsEnabledRef,
-    segment,
-    writer: {
-      writePreviewLineBreak: (outputSegment, onComplete) => {
-        writeLineBreakBeforePreview(handlers, outputSegment, onComplete)
-      },
-      writePreviews: (paths, onComplete) => {
-        writeInlineImagePreviews(handlers, paths, onComplete)
-      },
-      writeText: (text, onComplete) => {
-        handlers.terminal.write(text, onComplete)
-      }
-    }
-  }, () => {
-    handlers.lifecycle.outputWriting = false
-    flushTerminalOutputQueue(handlers)
-  })
-}
-
-const enqueueTerminalOutput = (
-  handlers: TerminalMessageHandlers,
-  data: string
-): void => {
-  for (const segment of splitTerminalInlineImageOutput(data)) {
-    handlers.lifecycle.outputQueue.push(segment)
-  }
-  flushTerminalOutputQueue(handlers)
-}
-
 const handleTerminalServerMessage = (
   handlers: TerminalMessageHandlers,
   payload: string
@@ -419,38 +201,6 @@ const attachTerminalSocketListeners = (
       onError(socket)
     }
   })
-}
-
-const closeSocket = (socket: WebSocket | null): void => {
-  if (socket === null || socket.readyState === WebSocket.CLOSED) {
-    return
-  }
-  runOptionalTerminalOperation(() => {
-    socket.close()
-  })
-}
-
-export const cleanupTerminalResources = (
-  args: TerminalCleanupArgs
-): void => {
-  args.lifecycle.disposed = true
-  clearReconnectTimer(args.lifecycle)
-  for (const disposable of args.lifecycle.inlineImageDisposables) {
-    disposable.dispose()
-  }
-  args.lifecycle.inlineImageDisposables = []
-  revokeTerminalInlineImageObjectUrlCache(args.lifecycle.inlineImageObjectUrls)
-  args.lifecycle.outputQueue = []
-  args.lifecycle.outputWriting = false
-  args.removeImageLinks()
-  args.removeImagePaste()
-  args.removeInput()
-  args.resizeObserver?.disconnect()
-  args.removeResize()
-  closeSocket(args.socketRef.current)
-  args.socketRef.current = null
-  args.runtimeRef.current = null
-  args.terminal.dispose()
 }
 
 const failBeforeAttach = (
